WebContent: signup: Add birthday and gender fields
Bythday is a field of type date when focused, It is of type text when
blur. We use this escamotage in order to show the placeholder and be
consistent with other fields.

For the gender we have three radio buttons in the same row-radio, the
display is flex without wrap on larger devices, with wrap on smaller.
The class that definies the style of the radio buttons is form-check,
that has a style very similar to form-control.

Change-Id: Ic4f4ddfc4a9ff59f425742722f833e06637ca94a